refactor(Http): split the nested classes into separate files
