+++ src/Google.Maps/Internal/Http.cs
@@ -16,9 +16,6 @@
  */
 
 using System;
-using System.IO;
-using System.Net;
-using Newtonsoft.Json;
 
 namespace Google.Maps.Internal
 {
@@ -27,61 +24,6 @@ namespace Google.Maps.Internal
 	/// </summary>
 	public static class Http
 	{
-		public class HttpGetResponse
-		{
-			protected Uri RequestUri { get; set; }
-
-			public HttpGetResponse(Uri uri)
-			{
-				RequestUri = uri;
-			}
-
-			protected virtual StreamReader GetStreamReader(Uri uri)
-			{
-				return GetStreamReader(uri, GoogleSigned.SigningInstance);
-			}
-			protected virtual StreamReader GetStreamReader(Uri uri, GoogleSigned signingInstance)
-			{
-				if(signingInstance != null)
-				{
-					uri = new Uri(signingInstance.GetSignedUri(uri));
-				}
-
-				WebResponse response = WebRequest.Create(uri).GetResponse();
-
-				StreamReader sr = new StreamReader(response.GetResponseStream());
-				return sr;
-			}
-
-			public virtual string AsString()
-			{
-				var output = String.Empty;
-
-				using(var reader = GetStreamReader(this.RequestUri))
-				{
-					output = reader.ReadToEnd();
-				}
-
-				return output;
-			}
-
-			public virtual T As<T>() where T : class
-			{
-				T output = null;
-
-				using(var reader = GetStreamReader(this.RequestUri))
-				{
-					JsonTextReader jsonReader = new JsonTextReader(reader);
-					JsonSerializer serializer = new JsonSerializer();
-					serializer.Converters.Add(new JsonEnumTypeConverter());
-					serializer.Converters.Add(new JsonLocationConverter());
-					output = serializer.Deserialize<T>(jsonReader);
-				}
-
-				return output;
-			}
-		}
-
 		public static HttpGetResponse Get(Uri uri)
 		{
 			return Factory.CreateResponse(uri);
@@ -91,22 +33,5 @@ namespace Google.Maps.Internal
 		/// Gets or sets the factory that provides HttpGetResponse instances. Crude depency injection for the time being.
 		/// </summary>
 		public static HttpGetResponseFactory Factory = new HttpGetResponseFactory();
-
-		/// <summary>
-		/// A factory class for building HttpGetResponse instances.
-		/// </summary>
-		public class HttpGetResponseFactory
-		{
-			/// <summary>
-			/// Builds a standard HttpGetResponse instance.
-			/// </summary>
-			/// <param name="uri"></param>
-			/// <returns></returns>
-			public virtual HttpGetResponse CreateResponse(Uri uri)
-			{
-				return new HttpGetResponse(uri);
-			}
-		}
-
 	}
 }
