+++ src/Google.Maps/Internal/HttpGetResponseFactory.cs
@@ -1,0 +1,37 @@
+﻿/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using System;
+
+namespace Google.Maps.Internal
+{
+	/// <summary>
+	/// A factory class for building HttpGetResponse instances.
+	/// </summary>
+	public class HttpGetResponseFactory
+	{
+		/// <summary>
+		/// Builds a standard HttpGetResponse instance.
+		/// </summary>
+		/// <param name="uri"></param>
+		/// <returns></returns>
+		public virtual HttpGetResponse CreateResponse(Uri uri)
+		{
+			return new HttpGetResponse(uri);
+		}
+	}
+}
